Refactor Ownership into a different struct
