+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tools/hdfs-chown/CMakeLists.txt
@@ -16,8 +16,8 @@
 # limitations under the License.
 #
 
-add_library(hdfs_chown_lib STATIC  $<TARGET_OBJECTS:hdfs_tool_obj> hdfs-chown.cc)
-target_include_directories(hdfs_chown_lib PRIVATE ../../tools hdfs-chown ${Boost_INCLUDE_DIRS})
+add_library(hdfs_chown_lib STATIC $<TARGET_OBJECTS:hdfs_tool_obj> $<TARGET_OBJECTS:hdfs_ownership_obj> hdfs-chown.cc)
+target_include_directories(hdfs_chown_lib PRIVATE ../../tools ${Boost_INCLUDE_DIRS})
 target_link_libraries(hdfs_chown_lib PRIVATE Boost::boost Boost::program_options tools_common hdfspp_static)
 
 add_executable(hdfs_chown main.cc)
