+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/tools/hdfs-ownership.cc
@@ -1,0 +1,44 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "hdfs-ownership.h"
+
+namespace hdfs::tools {
+Ownership::Ownership(const std::string &user_and_group) {
+  const auto owner_end = user_and_group.find(':');
+  if (owner_end == std::string::npos) {
+    user_ = user_and_group;
+    return;
+  }
+
+  user_ = user_and_group.substr(0, owner_end);
+  group_ = user_and_group.substr(owner_end + 1);
+}
+
+bool Ownership::operator==(const Ownership &other) const {
+  const auto same_user = user_ == other.user_;
+  if (group_.has_value() && other.group_.has_value()) {
+    return same_user && group_.value() == other.group_.value();
+  }
+
+  if (!group_.has_value() && !other.group_.has_value()) {
+    return same_user;
+  }
+  return false;
+}
+} // namespace hdfs::tools
